old/pyproject.toml
[tool.poetry]
name = "karbon-api-wrapper"
version = "0.1.0"
description = "API wrapper for Karbon data"
authors = ["Devin <devin@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.68.0"
uvicorn = "^0.15.0"
httpx = "^0.23.0"
python-dotenv = "^0.19.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
